pyproject.toml
[project]
name = "mcp-mysql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.9.2",
    "pymysql>=1.1.1",
    "python-dotenv>=1.0.0",
]
